fixed insertion of log retrieval of state
